move range_typet to mathematical_types.h
The header file mathematical_types.h is a better fit for range_typet, the
type for subranges of integers.
