+++ src/analyses/goto_rw.h
@@ -12,13 +12,15 @@ Date: April 2010
 #ifndef CPROVER_ANALYSES_GOTO_RW_H
 #define CPROVER_ANALYSES_GOTO_RW_H
 
-#include <iosfwd>
-#include <map>
-#include <memory> // unique_ptr
+#include <util/mp_arith.h>
+
+#include <goto-programs/goto_program.h>
 
 #include "guard.h"
 
-#include <goto-programs/goto_program.h>
+#include <iosfwd>
+#include <map>
+#include <memory> // unique_ptr
 
 class goto_functionst;
 class message_handlert;
